"light map" viz.
it's kinda cheat because this is not a light mapper but a "real-time" (6 FPS) path-tracer. At first hit, I'm injecting objColor to be white instead of albedo texture sample. For the rest I sample the texture. For light baking this cheat won't be needed because first hit won't be with the original surface, so this term won't be included in the first place